+++ src/com/actions/scene/launchar.ts
@@ -1,4 +1,4 @@
-import { type ARSystemOptions } from '../../../ar/ARSystem';
+import { type ARSystemOptions } from '../../../modules/ar/ARSystem';
 
 export default interface LAUNCH_AR {
     DESCRIPTION: 'Launches AR mode in native capabilities. (iOS: AR Quick Look, Android: Google Scene Viewer)';
