+++ src/modules/registry/__test__/ModuleRegistry.test.ts
